# Minimal dispatch