--- a/js/metrics.ts
+++ b/js/metrics.ts
@@ -34,7 +34,19 @@ function res(baseRes: null | msg.Base): Metrics {
   };
 }
 
-/** Receive metrics from the privileged side of Deno. */
+/** Receive metrics from the privileged side of Deno.
+ *
+ *    > console.table(Deno.metrics())
+ *   ┌──────────────────┬────────┐
+ *   │     (index)      │ Values │
+ *   ├──────────────────┼────────┤
+ *   │  opsDispatched   │   9    │
+ *   │   opsCompleted   │   9    │
+ *   │ bytesSentControl │  504   │
+ *   │  bytesSentData   │   0    │
+ *   │  bytesReceived   │  856   │
+ *   └──────────────────┴────────┘
+ */
 export function metrics(): Metrics {
   return res(dispatch.sendSync(...req()));
 }