--- a/src/startup_data.rs
+++ b/src/startup_data.rs
@@ -1,17 +1,8 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
 use deno_core::deno_buf;
+use deno_core::{StartupData, StartupScript};
 
-pub struct IsolateInitScript {
-  pub source: String,
-  pub filename: String,
-}
-
-pub struct IsolateInit {
-  pub snapshot: Option<deno_buf>,
-  pub init_script: Option<IsolateInitScript>,
-}
-
-pub fn deno_isolate_init() -> IsolateInit {
+pub fn deno_isolate_init() -> StartupData {
   if cfg!(feature = "no-snapshot-init") {
     debug!("Deno isolate init without snapshots.");
     #[cfg(not(feature = "check-only"))]
@@ -20,13 +11,10 @@ pub fn deno_isolate_init() -> IsolateInit {
     #[cfg(feature = "check-only")]
     let source_bytes = vec![];
 
-    IsolateInit {
-      snapshot: None,
-      init_script: Some(IsolateInitScript {
-        filename: "gen/bundle/main.js".to_string(),
-        source: std::str::from_utf8(source_bytes).unwrap().to_string(),
-      }),
-    }
+    StartupData::Script(StartupScript {
+      filename: "gen/bundle/main.js".to_string(),
+      source: std::str::from_utf8(source_bytes).unwrap().to_string(),
+    })
   } else {
     debug!("Deno isolate init with snapshots.");
     #[cfg(not(any(feature = "check-only", feature = "no-snapshot-init")))]
@@ -36,15 +24,12 @@ pub fn deno_isolate_init() -> IsolateInit {
     let data = vec![];
 
     unsafe {
-      IsolateInit {
-        snapshot: Some(deno_buf::from_raw_parts(data.as_ptr(), data.len())),
-        init_script: None,
-      }
+      StartupData::Snapshot(deno_buf::from_raw_parts(data.as_ptr(), data.len()))
     }
   }
 }
 
-pub fn compiler_isolate_init() -> IsolateInit {
+pub fn compiler_isolate_init() -> StartupData {
   if cfg!(feature = "no-snapshot-init") {
     debug!("Deno isolate init without snapshots.");
     #[cfg(not(feature = "check-only"))]
@@ -53,13 +38,10 @@ pub fn compiler_isolate_init() -> IsolateInit {
     #[cfg(feature = "check-only")]
     let source_bytes = vec![];
 
-    IsolateInit {
-      snapshot: None,
-      init_script: Some(IsolateInitScript {
-        filename: "gen/bundle/compiler.js".to_string(),
-        source: std::str::from_utf8(source_bytes).unwrap().to_string(),
-      }),
-    }
+    StartupData::Script(StartupScript {
+      filename: "gen/bundle/compiler.js".to_string(),
+      source: std::str::from_utf8(source_bytes).unwrap().to_string(),
+    })
   } else {
     debug!("Deno isolate init with snapshots.");
     #[cfg(not(any(feature = "check-only", feature = "no-snapshot-init")))]
@@ -69,10 +51,7 @@ pub fn compiler_isolate_init() -> IsolateInit {
     let data = vec![];
 
     unsafe {
-      IsolateInit {
-        snapshot: Some(deno_buf::from_raw_parts(data.as_ptr(), data.len())),
-        init_script: None,
-      }
+      StartupData::Snapshot(deno_buf::from_raw_parts(data.as_ptr(), data.len()))
     }
   }
 }