Perf improvements by switching to current thread runtime, and avoiding the use of flatbuffers for op_read and op_write. Sets the stage to abandon flatbuffers elsewhere.


master branch, `deno tests/http_bench.ts`
```
~/src/deno> third_party/wrk/mac/wrk -d 10 -c 10 http://127.0.0.1:4500/
Running 10s test @ http://127.0.0.1:4500/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.94ms   22.73ms 363.54ms   95.28%
    Req/Sec    17.93k     2.66k   22.92k    80.50%
  356819 requests in 10.00s, 17.35MB read
Requests/sec:  35676.99
Transfer/sec:      1.74MB
```

`node tools/node_tcp.js`
```
~/src/deno> third_party/wrk/mac/wrk -d 10 -c 10 http://127.0.0.1:4544/
Running 10s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   251.77us  123.99us   4.93ms   84.39%
    Req/Sec    19.51k     2.29k   26.57k    83.17%
  392022 requests in 10.10s, 19.07MB read
Requests/sec:  38815.10
Transfer/sec:      1.89MB
```

**this branch**, `deno tests/http_bench.ts`
```
~/src/deno> third_party/wrk/mac/wrk -d 10 -c 10 http://127.0.0.1:4500/
Running 10s test @ http://127.0.0.1:4500/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   525.63us    1.57ms  54.80ms   94.10%
    Req/Sec    24.10k     1.61k   27.45k    80.50%
  479501 requests in 10.00s, 23.32MB read
Requests/sec:  47935.58
Transfer/sec:      2.33MB
```

this branch, `deno_core_http_bench`
```
~/src/deno> third_party/wrk/mac/wrk -d 10 -c 10 http://127.0.0.1:4544/
Running 10s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   402.90us    1.15ms  20.76ms   94.86%
    Req/Sec    26.86k     2.01k   31.81k    78.71%
  539721 requests in 10.10s, 26.25MB read
Requests/sec:  53435.75
Transfer/sec:      2.60MB
```