@ry 2019-05-01
closed in favor of https://github.com/denoland/deno/pull/2259